rainfall_prediction/requirements.txt
Flask
scikit-learn
pandas
numpy
joblib
requests
python-dotenv
